% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots all accidents in a given state during a given year in a map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the index number of a state}

\item{year}{the number of a year}
}
\description{
This function takes the number of a state and the number of a year and plots
all accidents from the FARS data set which happened during the given year in
the given state. Each accident is represented by a point on a map of the
given state.
}
\details{
This function returns NULL, but as a side effect the state map as
described above is plotted.

This function throws an error if the given state number does not
occur in the FARS data set.

Errors may also be thrown by calls to the plotting functions, such as
\code{maps::map("state")}. This happens, for example when calling
\code{fars_map_state(2, 2013)}.

One reason for errors could be missing data in the requested data set. This
function implicitly sets missing longitudes to 900 and missing latitudes to
90. These values may be outside the bounds of the state map and therefore
cause plotting to throw errors.
}
\examples{
fars_map_state(1, 2013)
}
