% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a file into a data frame tbl.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{the full file name of the CSV file to read}
}
\value{
a data frame tbl containing the data in the file read
}
\description{
This function reads a given comma separated values (CSV) file from the
current working directory and creates a data frame tbl from its contents.
}
\details{
All messages during reading are being suppressed.

The function will throw an error if a file with the given file name does not exist
}
\examples{
\dontrun{
dat <- fars_read("accident_2013.csv.bz2")
}
}
