% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads FARS data sets for multiple years and creates a list of tibbles.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integers, one for each year for which you want to
load the FARS data}
}
\value{
a list of tibbles, one for each year given
}
\description{
This function reads FARS data sets for multiple years from the current
working directory and creates a list of tibbles. The list contains as many
elements as years were given. Each tibble has two columns: \code{MONTH}
and \code{year}, where \code{year} is the given year and \code{MONTH}
contains the number of the month to which each observation belongs.
}
\details{
This function uses \code{fars_read} and \code{make_filename}. These
functions may throw errors, for example, if files with names for the
requested FARS datasets do not exists in the current working directory.
}
\examples{
\dontrun{
fars_data <- fars_read_years(c(2013, 2014))
}
}
