% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Counts the number of observations per month for given years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of year numbers.}
}
\value{
a tibble with 12 rows and a number of columns equal to the number of
years given.
}
\description{
This function takes multiple year numbers, reads the FARS data sets for
the given years from the current working directory and counts the number
of observations per month for each year. The result is a tibble with a
number of columns equal to the number of years given and 12 rows, one for
each month. Each cell of the tibble contains the number of observations per
month and year.
}
\details{
This function uses \code{fars_read} and \code{make_filename}. These
functions may throw errors, for example, if files with names for the
requested FARS datasets do not exists in the current working directory.
}
\examples{
\dontrun{
y <- fars_summarize_years(c(2013, 2014))
}
}
