% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name from a given year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the number of the year for whose data set you want the file name}
}
\value{
the file name of the FARS data set for the given year
}
\description{
This function creates a string that is the file name for the FARS data set
of a given year.
}
\details{
The function will try to cast the parameter into an integer. Errors
may result from casting a variable that cannot be cast into an integer.
}
\examples{
f <- make_filename(2013)
}
